+++ cli/genoboo.js
@@ -849,6 +849,10 @@ addEggnog
   .description('Add EggNog-mapper results to a running GeneNoteBook server')
   .usage('[options] <EggNog-mapper tsv output file>')
   .arguments('<file>')
+  .option(
+    '--gofile <gofile>',
+    'Path to Gene Ontology json file',
+  )
   .requiredOption(
     '-u, --username <adminUsername>',
     'GeneNoteBook admin username'
@@ -865,7 +869,7 @@ addEggnog
     '--port [port]',
     'Port on which GeneNoteBook is running. Default: 3000'
   )
-  .action((file, { username, password, port = 3000, annot }) => {
+  .action((file, { username, password, port = 3000, annot, goFile }) => {
     if (typeof file !== 'string') addEggnog.help();
 
     const fileName = path.resolve(file);
@@ -875,7 +879,8 @@ addEggnog
 
     new GeneNoteBookConnection({ username, password, port }).call('addEggnog', {
       fileName,
-      annot: annot
+      annot: annot,
+      goFile
     });
   })
   .on('--help', () => {