+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
-## [0.4.19] TBA
+## [0.4.19] 2026-03-12
 
 - Properly raises error when failing to add an annotation (#82)
 - Fix eggnog seed link (#89)